+++ tests/testthat/test-nice_fit.R
@@ -28,7 +28,7 @@ test_that("nice_fit regular", {
 })
 
 test_that("nice_fit as nice_table", {
-   skip_if_not_installed("rempsyc")
+  skip_if_not_installed("rempsyc")
   expect_s3_class(
     nice_fit(fit, nice_table = TRUE), "flextable"
   )
@@ -72,14 +72,16 @@ test_that("nice_fit error on wronb object class", {
   )
 })
 
-dat <- data.frame(z = sample(c(0, 1), 100, replace = TRUE),
-                  x = sample(1:7, 100, replace = TRUE),
-                  y = sample(1:5, 100, replace = TRUE))
-mod <- '
+dat <- data.frame(
+  z = sample(c(0, 1), 100, replace = TRUE),
+  x = sample(1:7, 100, replace = TRUE),
+  y = sample(1:5, 100, replace = TRUE)
+)
+mod <- "
   y ~ a*x
   z ~ b*y + c*x
   ind := a*b
-  '
+  "
 fit <- sem(mod, dat, ordered = "z")
 
 test_that("nice_fit test categorical variable", {
@@ -92,5 +94,3 @@ test_that("nice_fit test categorical variable", {
     nice_fit(fit, nice_table = TRUE, verbose = FALSE), "flextable"
   )
 })
-
-